Move Refresh and Retention structs to AcceleratedTable
